Kernel: Implement memory regions and page fault handling.
We introduce memory regions and the abillity to allocate new pages in
virtual memory.

A region is a contigous range of memory that is
repsonsible for handling the memory permissions of that region;
furthermore, it handles any page fault that occurs in that region.
Currently, we create regions for each section of the kernel executable
and another for managing the kernel's heap.

Upon a "not-present" page fault, we can now allocate a physical page to
be assigned to the corresponding virtual address. At this time, we use a
water mark allocator, which should be replaced in the near future.