Kernel: fix bug in Thread::wait_for_cpu.
Previously, Thread::wait_for_cpu allocated a blocker on the stack; this
blocker would fall out of scope, allowing the blocker to become
corrupted.
This is fixed by instead passing the blocker to wait_for_cpu by
reference, providing us better control of its lifetime. This is kind of
a gross fix but it'll do for now.