Kernel: create userspace threads with Thread(File*)
We can now construct and run userspace threads by constructing a Thread
with a File. This involves setting up per-thread userspace regions and
the mechanisms to handle faults in these regions.

We've had to introduce PagingScope, which swaps into a threads page
directory. This is so that we can write the executable into the thread's
memory on construction. This'll need to be tidied up in the future, but
should be fine for now.

It's important to note that, for now, all pages are set as
userspace-accessible. This is terrible for security, but made
implementation a lot easier and should be easy enough to fix in the
future.