Kernel: fix indexing error in Paging::lookup_page_entry
Due to a bit masking error when calculating indices in lookup_page_entry,
on high addresses, we would perform an out of bounds access when reading
the entry from the page table.

This is fixed by changing the bitmask from 0xfff to 0x3ff. The rationale
for this is that a page table has 1024 entries -from 0 to 1023- and the
highest number representable by 10 bits is 1023, hence 0x3ff.