Kernel: add system calls.
We can now call system calls from userspace. Currently, the only
existing syscall is sys_debug, which prints a string to com1.
We use a bsd-style syscall mechanism where arguments are passed the the
kernel via the user stack, eax is used to specify the syscall number and
int 0x80 triggers the syscall mechanism.

At the moment we are very naive and do not verify user arguments; this
will have to change in the future.