TTY: Speed up scroll speed.
Previously, we were doing too much reading and writing from the terminals
vga memory, which is dramatically slower than standard memory. Now we
use an in memory buffer so that we can avoid vga reads and, in the case
that an update doesn't change anything, vga writes.

This dramatically improves scrolling speed.