Kernel: Add Thread::mark_for_death
This commit changes the way that threads die. We now use mark_for_death
to say that we want the thread to die when it returns to userspace.
Note that the thread will either be running in an interrupt or blocking
in the kernel, so we shouldn't have to wait long for the thread to
actually die.