TTY: yield after receiving a line.
We now yield whenever the tty receives a complete line.
This reduces latency, as previously we would have to
wait for the current thread's quantum to end before the waiting
thread could handle the input. This was particularly problemeatic
when the idle thread received the keyboard interrupt.

This works well at the moment, but will be even better when we
thread priorities.