Userspace: Store .bss in data region.
Because we use simple flat binaries, the .bss section
wasn't being loaded properly. This caused some
variables to cause segfaults, as they weren't included
in the binary.

To fix this we include all bss sections in .data. This is
very hacky, but'll do until we get a proper ELF loader.