Kernel+signals: don't setup signals when returning into kernel space.
When setting up signals, we did not check that we were returning into
userspace. This meant that, when handling nested interrupts (such as
page faults), we could start setting up a signal context in the middle
of another signal context, corrupting the programs state.

Also, we now store the signal trampoline in a per process region. This
is done because it is easier to implement than a stack based trampoline.