Thread: add basic ELF support.
We can now load and run simple elf files. The implementation
and api are currently kind of poor, but it's a start!

The main issues are:
Thread constructor takes both elf and flat files.
We allocate a region of all the program sections, this
is too trusting and makes an unnecessary allocation.
We don't handle BSS properly.